moved variations into 1 file
